styling and linters error fixed
