--- src/components/Calculator.css
@@ -1,5 +1,0 @@
-.calc-wrap {
-  position: fixed;
-  right: 250px;
-  top: 220px;
-}
