+++ src/components/Calculator/Calculator.js
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import calculate from '../logic/calculate';
+import calculate from '../../logic/calculate';
 import './Calculator.css';
 
 const Calculator = () => {
