+++ src/index.css
@@ -7,6 +7,7 @@ body {
   background-repeat: no-repeat;
   background-size: cover;
   font-family: 'Barlow', cursive;
+  height: auto;
 }
 
 :root {
